chore(ds): update utils
